users/delay: add support for platform-specific keycodes
Add user-specific settings that can be persisted in EEPROM.  Optionally
stores the target platform in user-specific settings, and changes the
behavior of cut/copy/paste keycodes based on the currently targetted
platform.

Add keycodes to change the target platform at runtime.  Use LEDs when
available to confirm the selected platform.